Qualification/pass ball (#102)
* Add qualification criteria 3 basics

* Add basic gotopos skill

* Add better PID control ish

* Remove penaltyarea for robots

* improve coach for defenders etc

* add more coach functions

* fix compiler

* use coach functions in defend skill

* ObjectAvoidance using luTh is working properly

* add to switches and clean switches a little

* fix switches params and add Defend in a json for testing

* re add inverter

* add tactic with two working defenders

* add some more functionality for defenders

* Fix pid gtpluth

* Change PID values

* add halt button

* add callback to interfaceValues

* Add halt functionality

* Call halt from interface

* fix interface segfault

* Add halt check

* Add warning to halt

* Comment out some failing tests

* zComitttmezasge

* zComitttmezasge

* Update Cmake

* Fix tests

* Fix compile warnings

* Improved GoToPosLuTh by a bunch

* baris is the best twee

* Qualification 3 done (ish)

* if it works it works

* baris is noob 3

* baris is retereedered

+++ roboteam_ai/src/conditions/IsRobotClosestToBall.cpp
@@ -30,7 +30,7 @@ namespace ai{
 
         if (robotClosestToBallPtr && robot) {
             robotClosestToBall = *robotClosestToBallPtr;
-            if (robot->id == robotClosestToBall) {
+            if (robot->id == static_cast<unsigned int>(robotClosestToBall)) {
                 return Status::Success;
             } else {
                 return Status::Failure;
